XHTML: Omit empty lists from output
(X)HTML lists are defined as containing one or more list
elements. Empty lists are not defined, so we omit the whole
list if it does not have any elements. This is the non-hackish
aspect of this patch.

However, just dropping the empty list might cause similar issues
with now suddenly empty enveloping elements, or break some CSS
based page layouts which rely on styling the omitted list. That
is the hackish aspect of this patch.