SC: Move external CSS content to SVG code
At least for now, having the CSS properties in line in
the SVG code makes more sense than having it in a separate
file.

Most of the styles will be defined via the #svg_chart
struct in the future, so having them generated along
with the SVG code makes sense.